% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{get_results}
\alias{get_results}
\title{Load and analyse YAPOS results}
\usage{
get_results(nc_output_file, create_plots = TRUE)
}
\arguments{
\item{nc_output_file}{Path of the NetCDF containing the results}

\item{create_plot}{Flag to define if the function should generate also the plots}
}
\value{
A list containing results' data and plots
}
\description{
This function loads all the results of the simulation contained in the target NetCDF file
and returns a set of pre-defined plots and tables
}
\details{
Returned data include the following data frames:

\itemize{
\item `all_gen`: details on generating units
\item `all_lines`: datailed list of transmission lines
\item `avail`: availability time-series
\item `final_dem`: time-series of demand
\item `ZONES`: vector containing all the names of the simulated zones
\item `final_ren`: time-series of non-dispatchable renewables
\item `shed`
\item `curt`
\item `sl`
\item `final_prod_detailed`
\item `compare_entsoe`
\item `final_co2`
\item `flow`
\item `summary_curt_shed`
\item `storage_level`
\item `all_inflow`
\item `dual`
}

The plots are instead:
\itemize{
\item `single_res`
\item `all_res`
\item `dispatch`
\item `annual_gen`
\item `entsoe`
\item `co2`
\item `flow`
\item `inflow`
\item `curtailment`
\item `shedding`
}

The dual data is loaded only if the function finds in the same directory of `nc_output_file`
the txt files containing the dual results saved by YAPOS.
}
\examples{
target_nc_file <- system.file("extdata", "es-pt-fr.nc", package = "yaposer", mustWork = TRUE)
res <- get_results(target_nc_file)
}
