% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_input_from_folder.R
\name{read_input_from_folder}
\alias{read_input_from_folder}
\title{Load YAPOS input data.}
\usage{
read_input_from_folder(folder)
}
\arguments{
\item{folder}{Path of the directory containing the files}
}
\value{
A list containing all the data
}
\description{
This function loads all the input data for a YAPOS simulation into a `list`
}
\details{
The function looks for in the specified folder the following CSV files:
\itemize{
  \item `gen.csv`: generation units
  \item `lin.csv`: tranmission lines
  \item `dem.csv`: demand time-series
  \item `ren.csv`: time-series of non-dispatchable renewables
  \item `inflow.csv`: time-series of inflow
  \item `avail.csv`: time-series of units' availabilities
  \item `stomin`: time-series with the storage minimum
}
}
