% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_input_from_netcdf.R
\name{read_input_from_netcdf}
\alias{read_input_from_netcdf}
\title{Read simulation data from a NetCDF file}
\usage{
read_input_from_netcdf(nc_output_file)
}
\arguments{
\item{nc_output_file}{Path to the NetCDF file}
}
\value{
A list with all the inputs
}
\description{
This function opens a NetCDF file containing the results of a YAPOS simulation and extracts the inputs.
}
\examples{
target_nc_file <- system.file("extdata", "es-pt-fr.nc", package = "yaposer", mustWork = TRUE)
inputs <- read_inputs_from_netcdf(target_nc_file)
}
