% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_sim.R
\name{subset_sim}
\alias{subset_sim}
\title{Subset simulation data}
\usage{
subset_sim(sim_data, zones)
}
\arguments{
\item{sim_data}{Simulation inputs loaded with `read_input_from_folder` or `read_input_from_netcdf`}

\item{zones}{Vector of the zones to be used for subsetting}
}
\value{
YAPOS input data
}
\description{
This function process YAPOS input data limiting the simulation to a list of zones.
}
\examples{
target_nc_file <- system.file("extdata", "es-pt-fr.nc", package = "yaposer", mustWork = TRUE)
input_data_3_zones <- read_input_from_netcdf(target_nc_file)
input_data_2_zones <- subset_sim(input_data_3_zones, zones = c('ES', 'PT'))
}
