% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_results.R
\name{viz_results}
\alias{viz_results}
\title{Visualise simulation results}
\usage{
viz_results(res)
}
\arguments{
\item{res}{Results loaded with `get_results`}
}
\description{
This function launches a Shiny application able to visualize and explore simulation data.
}
\details{
Simulation data must be loaded using `get_results`
}
