% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sim_data.R
\name{write_sim_data}
\alias{write_sim_data}
\title{Write simulation data}
\usage{
write_sim_data(sim_data, target_folder)
}
\arguments{
\item{sim_data}{Simulation inputs loaded with `read_input_from_folder` or `read_input_from_netcdf`}

\item{target_folder}{Target folder}
}
\description{
this function writes the input data into a folder for a YAPOS simulation.
}
