% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input_data.R
\name{check_input_data}
\alias{check_input_data}
\title{Check if YAPOS input data is well-formed and return also pre-processing analysis.}
\usage{
check_input_data(s)
}
\arguments{
\item{s}{List containing all}
}
\value{
A list containing all the data
}
\description{
This function check if a list contains all the right data structures needed for
a YAPOS simulation. It also checks if they are well-formed (dimensions and consistency).
}
\details{
The function returs also a data frame containing pre-processing information on the simulation
}
