% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{get_results}
\alias{get_results}
\title{Load and analyse YAPOS results}
\usage{
get_results(nc_output_file, create_plots = TRUE)
}
\arguments{
\item{nc_output_file}{Path of the NetCDF containing the results}

\item{create_plot}{Flag to define if the function should generate also the plots}
}
\value{
A list containing results' data and plots
}
\description{
This function loads all the results of the simulation contained in the target NetCDF file
and returns a set of pre-defined plots and tables. If the text output files containing the simulation dual
values are present in the same folder of the NetCDF file, this function loads and store their value in a data frame.
}
\details{
Returned data include the following data frames:

\itemize{
\item `all_gen`: details on generating units
\item `all_lines`: datailed list of transmission lines
\item `avail`: availability time-series
\item `final_dem`: time-series of demand
\item `ZONES`: vector containing all the names of the simulated zones
\item `final_ren`: time-series of non-dispatchable renewables
\item `shed`: time-series of daily shed load
\item `curt`: time-series of daily curtailed electricity
\item `storage_level`: storage levels at daily resolution for each generating unit. This data frame contains also availability and storage minimum
\item `final_prod_detailed`: generation for each generation unit including fuel type, zone of the unit and the emissions per MWh
\item `compare_entsoe`: comparison of the annual generation per zone and fuel with the ENTSO-E data (Statistical Factsheets) for 2016
\item `final_co2`: time-series of CO2 generation per zone
\item `flow`: time-series of electricity flow per line. It contains also the normalised values.
\item `summary_curt_shed`: table with the annual values of shed load and curtailed electricity
\item `all_inflow`: cumulated inflow per zone
\item `dual`: dual values for all problem constraints
}

The plots are instead:
\itemize{
\item `single_res`: storage level per single unit
\item `all_res`: area plot of cumulated storage levels
\item `dispatch`: dispatch plots per zone
\item `annual_gen`: summary of annual generation
\item `entsoe`: comparison with ENTSO-E
\item `co2`: area plot of the cumulated CO2
\item `flow`: heat-map with the electricity normalised flows
\item `inflow`: cumulated inflows
\item `curtailment`: plot of curtailment
\item `shedding`: plot of shedding load
}

The dual data is loaded only if the function finds in the same directory of `nc_output_file`
the txt files containing the dual results saved by YAPOS.
}
\examples{
target_nc_file <- system.file("extdata", "es-pt-fr.nc", package = "yaposer", mustWork = TRUE)
res <- get_results(target_nc_file)
}
